cellsLabels merged into CellsHash
