+++ src/gridGL/cells/cellsLabel/bitmapTextUtils.ts
